fix: comprehensive code review fixes for enterprise-claw-4j
Apply all HIGH/MEDIUM/LOW severity fixes identified during code audit:

HIGH:
- H1: AgentLoop now passes system prompt to Claude API via PromptAssembler
- H2: AgentLoop persists tool_use/tool_result events to JSONL
- H3: GatewayWebSocketHandler uses constructor injection
- H4: CronJob/CronPayload adds Jackson annotations for serialization
- H5: DeliveryQueue connects exponential backoff retry logic

MEDIUM:
- M1: FeishuChannel adds recursion depth guard
- M2: FeishuWebhookController adds HMAC-SHA256 signature verification
- M3: SessionStore uses full UUID with collision check
- M4: SessionStore loadSession reads under agent-level lock
- M5: MemoryStore thread safety (ConcurrentHashMap + volatile)
- M6: MemoryStore TF-IDF vector dimension fix
- M7: ContextGuard break->continue in Stage 2
- M8: ResilienceRunner model degradation fallback
- M9: BindingStore/AgentStore file locking
- M10: GatewayController proper getSessionHistory
- M11: Deduplicate send logic via GatewayService
- M12: TelegramChannel Markdown escaping
- M13: MessagePumpService @PreDestroy for clean shutdown

LOW:
- L1: TokenUsage int→long
- L2: ChannelManager stopReceiving/closeAll semantic separation
- L5: BindingTable cached sorted list
- L6: GlobalExceptionHandler hides internal errors in production
- L7: WebSocketConfig CORS configurable
- L8: AuthProfile caches AnthropicClient
- L9: HeartbeatService SpEL for ms conversion
- L10: AgentManager.normalizeAgentId()

New files: GatewayService.java, CronPayloadDeserializer.java
Total: 26 files changed, 584 insertions, 176 deletions

+++ claw-4j/enterprise-claw-4j/src/main/java/com/openclaw/enterprise/agent/TokenUsage.java
@@ -13,6 +13,6 @@ package com.openclaw.enterprise.agent;
  * @param outputTokens 输出 token 数 (Claude 生成的文本 + 工具调用)
  */
 public record TokenUsage(
-    int inputTokens,
-    int outputTokens
+    long inputTokens,
+    long outputTokens
 ) {}
